+++ contrib/opbot/botmd/commands.go
@@ -6,6 +6,14 @@ import (
 	"context"
 	"errors"
 	"fmt"
+	"log"
+	"math/big"
+	"regexp"
+	"strings"
+	"sync"
+	"time"
+
+	"github.com/dustin/go-humanize"
 	"github.com/ethereum/go-ethereum/accounts/abi/bind"
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/core/types"
@@ -17,12 +25,6 @@ import (
 	rfqClient "github.com/synapsecns/sanguine/services/rfq/api/client"
 	"github.com/synapsecns/sanguine/services/rfq/contracts/fastbridge"
 	"github.com/synapsecns/sanguine/services/rfq/relayer/relapi"
-	"log"
-	"math/big"
-	"regexp"
-	"strings"
-	"sync"
-	"time"
 )
 
 func (b *Bot) requiresSignoz(definition *slacker.CommandDefinition) *slacker.CommandDefinition {
@@ -138,6 +140,7 @@ func (b *Bot) traceCommand() *slacker.CommandDefinition {
 	})
 }
 
+// nolint: gocognit
 func (b *Bot) rfqLookupCommand() *slacker.CommandDefinition {
 	return &slacker.CommandDefinition{
 		Command:     "rfq <tx>",
@@ -204,7 +207,13 @@ func (b *Bot) rfqLookupCommand() *slacker.CommandDefinition {
 			}
 
 			var slackBlocks []slack.Block
+
 			for _, status := range statuses {
+				time, err := b.getTxAge(ctx, status.GetQuoteRequestStatusResponse)
+				if err != nil {
+					log.Printf("error getting tx age: %v\n", err)
+				}
+
 				objects := []*slack.TextBlockObject{
 					{
 						Type: slack.MarkdownType,
@@ -221,6 +230,10 @@ func (b *Bot) rfqLookupCommand() *slacker.CommandDefinition {
 					{
 						Type: slack.MarkdownType,
 						Text: fmt.Sprintf("*OriginTxHash*: %s", toTXSlackLink(status.OriginTxHash, status.OriginChainID)),
+					},
+					{
+						Type: slack.MarkdownType,
+						Text: fmt.Sprintf("*Estimated Tx Age*: %s", humanize.Time(time)),
 					},
 				}
 
@@ -332,6 +345,25 @@ func (b *Bot) makeFastBridge(ctx context.Context, req *relapi.GetQuoteRequestRes
 		return nil, fmt.Errorf("error creating fast bridge: %w", err)
 	}
 	return fastBridgeHandle, nil
+}
+
+func (b *Bot) getTxAge(ctx *slacker.CommandContext, status *relapi.GetQuoteRequestStatusResponse) (time.Time, error) {
+	// TODO: add CreatedAt field to GetQuoteRequestStatusResponse so we don't need to make network calls?
+	client, err := b.rpcClient.GetChainClient(ctx.Context(), int(status.OriginChainID))
+	if err != nil {
+		return time.Time{}, fmt.Errorf("error getting chain client: %w", err)
+	}
+
+	receipt, err := client.TransactionReceipt(ctx.Context(), common.HexToHash(status.OriginTxHash))
+	if err != nil {
+		return time.Time{}, fmt.Errorf("error fetching transaction receipt: %w", err)
+	}
+	txBlock, err := client.BlockByHash(ctx.Context(), receipt.BlockHash)
+	if err != nil {
+		return time.Time{}, fmt.Errorf("error fetching block by hash: %w", err)
+	}
+
+	return time.Unix(int64(txBlock.Time()), 0), nil
 }
 
 func toExplorerSlackLink(ogHash string) string {