+++ core/metrics/spanutils.go
@@ -1,6 +1,9 @@
 package metrics
 
-import "go.opentelemetry.io/otel/trace"
+import (
+	"go.opentelemetry.io/otel/attribute"
+	"go.opentelemetry.io/otel/trace"
+)
 
 // EndSpanWithErr ends a span and records an error if one is present.
 func EndSpanWithErr(span trace.Span, err error) {
@@ -9,6 +12,7 @@ func EndSpanWithErr(span trace.Span, err error) {
 	}
 
 	if err != nil {
+		span.SetAttributes(attribute.String("span_error", err.Error()))
 		span.RecordError(err)
 	}