+++ packages/widget/src/types/index.d.ts
@@ -84,6 +84,13 @@ export interface Chain {
   explorerName: string
   blockTime: number
   imgUrl: string
+  networkName: string
+  networkUrl: string
+  nativeCurrency: {
+    name: string
+    symbol: string
+    decimals: number
+  }
 }
 
 export declare function Bridge(props: BridgeProps): JSX.Element