+++ services/rfq/relayer/relapi/server.go
@@ -101,6 +101,7 @@ const (
 	getQuoteStatusByTxIDRoute   = "/status/by_tx_id"
 	getRetryRoute               = "/retry"
 	postWithdrawRoute           = "/withdraw"
+	getTxHashByNonceRoute       = "/tx_hash/by_nonce"
 	getRequestByTxID            = "/request/by_tx_id"
 )
 
@@ -123,6 +124,7 @@ func (r *RelayerAPIServer) Run(ctx context.Context) error {
 
 	if r.cfg.EnableAPIWithdrawals {
 		engine.POST(postWithdrawRoute, h.Withdraw)
+		engine.GET(getTxHashByNonceRoute, h.GetTxHashByNonce)
 	}
 
 	r.engine = engine