+++ services/rfq/relayer/relapi/server_test.go
@@ -4,14 +4,13 @@ import (
 	"context"
 	"encoding/json"
 	"fmt"
-	"github.com/ethereum/go-ethereum/accounts/abi/bind"
-	"github.com/stretchr/testify/assert"
-	"github.com/synapsecns/sanguine/services/rfq/relayer/chain"
-	"github.com/synapsecns/sanguine/services/rfq/relayer/relconfig"
 	"math/big"
 	"net/http"
 	"testing"
 	"time"
+
+	"github.com/stretchr/testify/assert"
+	"github.com/synapsecns/sanguine/services/rfq/relayer/relconfig"
 
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/common/hexutil"
@@ -208,76 +207,6 @@ func (c *RelayerServerSuite) getTestQuoteRequest(status reldb.QuoteRequestStatus
 	}
 }
 
-func (c *RelayerClientSuite) TestEthWithdraw() {
-	backend := c.underlying.testBackends[uint64(c.underlying.originChainID)]
-
-	startBalance, err := backend.BalanceAt(c.GetTestContext(), testWithdrawalAddress, nil)
-	c.Require().NoError(err)
-
-	withdrawalAmount := big.NewInt(50)
-
-	_, err = c.Client.Withdraw(c.GetTestContext(), &relapi.WithdrawRequest{
-		ChainID:      uint32(backend.GetChainID()),
-		To:           testWithdrawalAddress,
-		Amount:       withdrawalAmount.String(),
-		TokenAddress: chain.EthAddress,
-	})
-	c.Require().NoError(err)
-
-	// Wait for the transaction to be mined
-	err = retry.WithBackoff(c.GetTestContext(), func(ctx context.Context) error {
-		balance, err := backend.BalanceAt(ctx, testWithdrawalAddress, nil)
-		if err != nil {
-			return fmt.Errorf("could not fetch balance %w", err)
-		}
-
-		expectedBalance := new(big.Int).Add(startBalance, withdrawalAmount)
-
-		if balance.Cmp(expectedBalance) != 0 {
-			return fmt.Errorf("balance not updated")
-		}
-
-		return nil
-	})
-	c.Require().NoError(err)
-}
-
-func (c *RelayerClientSuite) TestERC20Withdraw() {
-	backend := c.underlying.testBackends[uint64(c.underlying.originChainID)]
-
-	_, erc20 := c.underlying.deployManager.GetMockERC20(c.GetTestContext(), backend)
-
-	startBalance, err := erc20.BalanceOf(&bind.CallOpts{Context: c.GetTestContext()}, testWithdrawalAddress)
-	c.Require().NoError(err)
-
-	withdrawalAmount := big.NewInt(50)
-
-	_, err = c.Client.Withdraw(c.GetTestContext(), &relapi.WithdrawRequest{
-		ChainID:      uint32(backend.GetChainID()),
-		To:           testWithdrawalAddress,
-		Amount:       withdrawalAmount.String(),
-		TokenAddress: erc20.Address(),
-	})
-	c.Require().NoError(err)
-
-	// Wait for the transaction to be mined
-	err = retry.WithBackoff(c.GetTestContext(), func(ctx context.Context) error {
-		balance, err := erc20.BalanceOf(&bind.CallOpts{Context: ctx}, testWithdrawalAddress)
-		if err != nil {
-			return fmt.Errorf("could not get balance %w", err)
-		}
-
-		expectedBalance := new(big.Int).Add(startBalance, withdrawalAmount)
-
-		if balance.Cmp(expectedBalance) != 0 {
-			return fmt.Errorf("balance not updated")
-		}
-
-		return nil
-	})
-	c.Require().NoError(err)
-}
-
 func TestTokenIDExists(t *testing.T) {
 	cfg := relconfig.Config{
 		QuotableTokens: map[string][]string{
@@ -359,74 +288,4 @@ func TestToAddressIsWhitelisted(t *testing.T) {
 			assert.Equal(t, tt.expected, result)
 		})
 	}
-}
-
-func (c *RelayerClientSuite) TestEthWithdrawCLI() {
-	res, err := c.Client.Withdraw(c.GetTestContext(), &relapi.WithdrawRequest{
-		ChainID:      c.underlying.originChainID,
-		To:           common.HexToAddress(testWithdrawalAddress.String()),
-		Amount:       "1000000000000000000",
-		TokenAddress: chain.EthAddress,
-	})
-	c.Require().NoError(err)
-
-	// Wait for the transaction to be mined
-	err = retry.WithBackoff(c.GetTestContext(), func(ctx context.Context) error {
-		status, err := c.underlying.database.SubmitterDB().
-			GetNonceStatus(
-				ctx,
-				c.underlying.wallet.Address(),
-				big.NewInt(int64(c.underlying.originChainID)),
-				res.Nonce,
-			)
-		if err != nil {
-			return fmt.Errorf("could not get status %w", err)
-		}
-
-		if status != submitterdb.Stored {
-			return fmt.Errorf("transaction not mined")
-		}
-
-		return nil
-	})
-	c.Require().NoError(err)
-	c.Require().NotNil(res)
-}
-
-func (c *RelayerClientSuite) TestERC20WithdrawCLI() {
-	backend := c.underlying.testBackends[uint64(c.underlying.originChainID)]
-
-	_, erc20 := c.underlying.deployManager.GetMockERC20(c.GetTestContext(), backend)
-
-	startBalance, err := erc20.BalanceOf(&bind.CallOpts{Context: c.GetTestContext()}, testWithdrawalAddress)
-	c.Require().NoError(err)
-
-	withdrawalAmount := big.NewInt(1000000000000000000)
-
-	res, err := c.Client.Withdraw(c.GetTestContext(), &relapi.WithdrawRequest{
-		ChainID:      c.underlying.originChainID,
-		To:           common.HexToAddress(testWithdrawalAddress.String()),
-		Amount:       withdrawalAmount.String(),
-		TokenAddress: erc20.Address(),
-	})
-	c.Require().NoError(err)
-
-	// Wait for the transaction to be mined
-	err = retry.WithBackoff(c.GetTestContext(), func(ctx context.Context) error {
-		balance, err := erc20.BalanceOf(&bind.CallOpts{Context: ctx}, testWithdrawalAddress)
-		if err != nil {
-			return fmt.Errorf("could not fetch balance %w", err)
-		}
-
-		expectedBalance := new(big.Int).Add(startBalance, withdrawalAmount)
-
-		if balance.Cmp(expectedBalance) != 0 {
-			return fmt.Errorf("balance not updated")
-		}
-
-		return nil
-	})
-
-	c.Require().NoError(err)
-	c.Require().NotNil(res)
 }