+++ services/rfq/relayer/service/handlers.go
@@ -20,7 +20,7 @@ import (
 	"go.opentelemetry.io/otel/trace"
 )
 
-var maxRPCRetryTime = 15 * time.Second
+var maxRPCRetryTime = 30 * time.Second
 
 // handleBridgeRequestedLog handles the BridgeRequestedLog event.
 // Step 1: Seen
@@ -332,13 +332,14 @@ func (r *Relayer) handleRelayLog(ctx context.Context, req *fastbridge.FastBridge
 	}
 
 	// TODO: this can still get re-orged
-	err = r.db.UpdateQuoteRequestStatus(ctx, req.TransactionId, reldb.RelayCompleted, nil)
-	if err != nil {
-		return fmt.Errorf("could not update request status: %w", err)
-	}
 	err = r.db.UpdateDestTxHash(ctx, req.TransactionId, req.Raw.TxHash)
 	if err != nil {
 		return fmt.Errorf("could not update dest tx hash: %w", err)
+	}
+
+	err = r.db.UpdateQuoteRequestStatus(ctx, req.TransactionId, reldb.RelayCompleted, nil)
+	if err != nil {
+		return fmt.Errorf("could not update request status: %w", err)
 	}
 	return nil
 }
@@ -373,6 +374,10 @@ func (q *QuoteRequestHandler) handleRelayCompleted(ctx context.Context, _ trace.
 //
 // This is the seventh step in the bridge process. Here we process the event that the proof was posted on chain.
 func (r *Relayer) handleProofProvided(ctx context.Context, req *fastbridge.FastBridgeBridgeProofProvided) (err error) {
+	if req.Relayer != r.signer.Address() {
+		return nil
+	}
+
 	// TODO: this can still get re-orged
 	// ALso: we should make sure the previous status  is ProvePosting
 	err = r.db.UpdateQuoteRequestStatus(ctx, req.TransactionId, reldb.ProvePosted, nil)